nsu-audit-mcp/pyproject.toml
[project]
name = "nsu-audit-mcp"
version = "2.0.0"
description = "NSU Audit Server - Transcript processing, certificate generation, and audit services"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "PyJWT>=2.8.0",
    "google-auth>=2.20.0",
    "google-auth-oauthlib>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
nsu-audit = "server:main"
nsu-audit-server = "server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["services", "models", "tools"]